refactor(api): only use Expectation as an Entity.
* it it not worth the time, the care about all those relationships.
* maybe we won't stick with mssql

+++ src/NinjaTools.FluentMockServer.Domain/Models/ValueTypes/TimeUnit.cs
@@ -9,7 +9,7 @@ namespace NinjaTools.FluentMockServer.Domain.Models.ValueTypes
     /// </summary>
     [JsonConverter(typeof(ContractResolver.UpperCaseEnumConverter))]
     [Serializable]
-    public enum TimeUnit : ushort
+    public enum TimeUnit : int
     {
         /// <summary>
         ///     The nanoseconds.
